requirements.txt
argparse
gitpython
numpy
pillow
halo